+++ Amazon/Pay/API/Constants/EvidenceType.php
@@ -8,5 +8,9 @@ class EvidenceType {
     public const CANCELLATION_POLICY = "CancellationPolicy";
     public const CUSTOMER_SIGNATURE = "CustomerSignature";
     public const TRACKING_NUMBER = "TrackingNumber";
+    public const CARRIER_NAME= "CarrierName";
+    public const DEVICE_ID= "DeviceId";
+    public const DEVICE_NAME= "DeviceName";
+    public const DOWNLOAD_DATE_TIME= "DownloadDateTime";
     public const OTHER = "Other";
 }